+++ extensions/subscriptions/js/src/@types/shims.d.ts
@@ -1,0 +1,7 @@
+import 'flarum/common/models/Discussion';
+
+declare module 'flarum/common/models/Discussion' {
+  export default interface Discussion {
+    subscription(): string;
+  }
+}
